fix: canvas table stability for multiple table instances
